refactor #441: ProfileScreen 컴포저블 파일 분리
